Move common IPvXAddressHelper.cs members from Net.Primitives and Private.Uri to Common (#33495)
* move common IPvXAddressHelper.cs pieces to Common

* Fix System.Net.Primitives.Pal.Tests

(cherry picked from commit 8280ae034fc1b2dc6f08a40afef0ddb4662d35a2)
